Fix: For bug "instances disabled by using the c++ api can not enabled via the command line or the config file"
Do not _process_requirements() for subcommands disabled or not required.
Do not use CLI11 disabled() to disable an instance. Set to not required is enough.
Do not require an instance in case the instance is disabled
Fix needs handling of instances so that setDisabled() and set a required socket address part (e.g. hostname, port) can be called in any order.